+++ src/pto/sk_pto_workflow_run.c
@@ -54,9 +54,9 @@ int _module_run(sk_sched_t* sched, sk_sched_t* src,
     SK_ENV_POS = SK_ENV_POS_CORE;
 
     if (ret) {
-        sk_print("module (%s) encounter error\n", module->cfg->name);
+        sk_print("module (%s) encounter error\n", module_name);
         SK_LOG_ERROR(SK_ENV_LOGGER,
-            "module (%s) encounter error", module->cfg->name);
+            "module (%s) encounter error", module_name);
 
         sk_txn_setstate(txn, SK_TXN_ERROR);
     }
@@ -71,18 +71,18 @@ int _module_run(sk_sched_t* sched, sk_sched_t* src,
         }
 
         sk_print("txn pending, waiting for service io calls, module %s\n",
-                 module->cfg->name);
+                 module_name);
         SK_LOG_TRACE(SK_ENV_LOGGER, "txn pending, waiting for service calls, "
-                     "module %s", module->cfg->name);
+                     "module %s", module_name);
         return 0;
     }
 
     if (ret) {
         sk_print("module (%s) encounter error, goto pack directly\n",
-                 module->cfg->name);
+                 module_name);
 
         SK_LOG_ERROR(SK_ENV_LOGGER,
-            "module (%s) encounter error, goto pack directly", module->cfg->name);
+            "module (%s) encounter error, goto pack directly", module_name);
         return _run(sched, src, entity, txn, proto_msg);
     }