+++ ChangeLog.md
@@ -1,6 +1,6 @@
 * 2016-08-15 0.9.1
   * Core
-    * **Fix:** Do not calling service api/job if service busy
+    * **Fix:** Do not calling service api/job if service is busy
     * **Fix:** Fix a potential memleak in txn.iocall if callback is NULL
     * **Enhancement:** Add an optional service config item for controlling max queue size
   * User